pyproject.toml
[project]
name = "ros-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "numpy>=2.4.4",
    "pydantic>=2.13.3",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.12",
    "scipy>=1.17.1",
    "uvicorn>=0.46.0",
]
